Create calenderConverter.py
Created calenderConverter.py, I know it is a mess, but it works. This will create an .ics file for each con using the JSON file as a guide, making creating calendar events a littler easier. The downside is you must update the range and the JSON file for each new conference date and info released.